liveSynth - avoid having to explicitly create synthdef
The defaults passed to the sd function are only used when first sending the synthdef. For creating the synth the params var is used. The default values for the sd function can therefore be derived from params by taking their value on the first tick. This ensures they will not change further which will avoid sending the synthdef again. Only changing sdbody will cause the synthdef to be sent again to scsynth.